requirements_test.txt
pytest>=7.0
pytest-asyncio>=0.21
pytest-homeassistant-custom-component>=0.13
pysolarmanv5>=3.0.0
